Title: 시력 검사 2, Time: 125ms, Memory: 24MB, Status: Passed - CodeTree

+++ 240110/시력 검사 2/eye-test-2.py
@@ -3,8 +3,8 @@ import sys
 input = sys.stdin.readline
 
 def solution():
-    a = int(input())
-    if a >= 1:
+    a = float(input())
+    if a >= 1.0:
         return 'High'
     elif a >= 0.5:
         return 'Middle'
